refactor: customer관련 dto 디렉토리 변경
