+++ src/main/java/woowacourse/shoppingcart/dto/customer/DeleteCustomerRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.customer;
 
 public class DeleteCustomerRequest {
     private String password;
