+++ src/test/java/woowacourse/shoppingcart/application/CustomerServiceTest.java
@@ -12,7 +12,7 @@ import org.springframework.test.context.jdbc.Sql;
 import woowacourse.common.exception.AuthException;
 import woowacourse.common.exception.JoinException;
 import woowacourse.shoppingcart.dao.CustomerDao;
-import woowacourse.shoppingcart.dto.CustomerResponse;
+import woowacourse.shoppingcart.dto.customer.CustomerResponse;
 
 
 @SpringBootTest
